// apps/web/src/api/__generated__/types/ActivePromotionResponseListItem.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * sync
 * OpenAPI spec version: 0.0.1
 */
import type { ActivePromotionResponseListItemFieldsItem } from './ActivePromotionResponseListItemFieldsItem';
import type { ActivePromotionResponseListItemSignup } from './ActivePromotionResponseListItemSignup';

export type ActivePromotionResponseListItem = {
  /**
   * 연결된 게시글의 슬러그
   * @nullable
   */
  postSlug?: string | null;
  /** 프로모션 ID */
  id: number;
  /** 신청 시 수집하는 필드 목록 */
  fields: ActivePromotionResponseListItemFieldsItem[];
  /** 내 신청 정보 */
  signup?: ActivePromotionResponseListItemSignup;
  /**
   * 연결된 게시글이 속한 프로젝트 핸들
   * @nullable
   */
  projectHandle?: string | null;
};