// apps/web/src/api/__generated__/types/ActivePromotionResponseListItemFieldsItemType.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * sync
 * OpenAPI spec version: 0.0.1
 */

/**
 * 필드 입력 타입
 */
export type ActivePromotionResponseListItemFieldsItemType =
  (typeof ActivePromotionResponseListItemFieldsItemType)[keyof typeof ActivePromotionResponseListItemFieldsItemType];

export const ActivePromotionResponseListItemFieldsItemType = {
  Text: 'TEXT',
  PhoneNumber: 'PHONE_NUMBER',
  Checkbox: 'CHECKBOX',
} as const;